Implemented faces settings dialog
Implemented faces settings dialog. Morph from spinbox to doublespinbox.

+++ dialogs/controllers/faces_settings_ctrl.hpp
@@ -12,7 +12,7 @@ class FacesSettingsCtrl : public QDialog
  public:
     explicit FacesSettingsCtrl(QWidget* parent=0);
 
-    void change_scale(int);
+    void change_scale(double);
     void change_min_neighbours(int);
 
  private:
